Fix sound mixing

+++ include/GameWindow.hpp
@@ -18,7 +18,7 @@ constexpr int SCREEN_TICKS_PER_FRAME = 1000 / SCREEN_FPS;
 class GameWindow
 {
 public:
-    GameWindow(NES& nes, char* frameBuffer, float* audioBuffer);
+    GameWindow(NES& nes, char* frameBuffer, int16_t* audioBuffer);
     ~GameWindow() = default;
     void Run();
 
@@ -27,7 +27,7 @@ public:
 private:
     NES& nes_;
     char* frameBuffer_;
-    float* audioBuffer_;
+    int16_t* audioBuffer_;
     SDL_AudioDeviceID audioDevice_;
 };
